Add files via upload
This branch represents the code which was utilized for the security function of encrypting the picture storage folder on the raspberry pi. While testing it was observed that while the application is running the user will not be able to input new pictures into the folder without crashing the application. and if the application is encrypted when not in use the user will not be able to input new pictures for each student. So due to these reasons we decided to incorporate the full access privilege security on the raspberry pi instead as this approach will not hinder the systems main processes.